ForestGeoProfile

Generated by interacting with app or function RequiredFormat_interactive

Usage

ForestGeoProfile

Format

A named list

TreeHeightUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your tree height unit?

Family

Value or column name in data set @ForestGeoSubset (Family) corresponding to Family name

TreeHeight

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree height

Date

Value or column name in data set @ForestGeoSubset (ExactDate) corresponding to Date of measurement

Month

Value or column name in data set @ForestGeoSubset (none) corresponding to Month of measurement

Day

Value or column name in data set @ForestGeoSubset (none) corresponding to Day of measurement

POM

Value or column name in data set @ForestGeoSubset (none) corresponding to POM

HOM

Value or column name in data set @ForestGeoSubset (hom) corresponding to HOM

HOMMan

Value or column name in data set @ForestGeoSubset (-999) corresponding to What is the height of diameter measurement in your inventory?

PlotAreaUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is the unit for plot area?

XTreeUTM

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree X UTM coordinate

IdStem

Value or column name in data set @ForestGeoSubset (stemID) corresponding to Stem ID

PlotMan

Value or column name in data set @ForestGeoSubset (ForestGEO) corresponding to What is your plot name ?

Species

Value or column name in data set @ForestGeoSubset (Species) corresponding to Species name

ScientificNameSepMan

Value or column name in data set @ForestGeoSubset () corresponding to Character separating Genus and Species in Scientific name

YTreePlot

Value or column name in data set @ForestGeoSubset (gy) corresponding to Tree Y euclidean position in plot

YTreeSubplot

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree Y euclidean position in SUBplot

CircUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your circumference unit?

BD

Value or column name in data set @ForestGeoSubset (none) corresponding to BD

BDUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your BD unit ?

BPOM

Value or column name in data set @ForestGeoSubset (none) corresponding to Basal POM

BHOM

Value or column name in data set @ForestGeoSubset (none) corresponding to Basal HOM

BHOMUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your BasalHOM unit?

BCirc

Value or column name in data set @ForestGeoSubset (none) corresponding to Basal Circumference

BCircUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your basal circumference unit?

IdMeasure

Value or column name in data set @ForestGeoSubset (MeasureID) corresponding to Measure ID

IdLevel

Value or column name in data set @ForestGeoSubset (IDLevel) corresponding to Determination rank

TreeFieldNum

Value or column name in data set @ForestGeoSubset (tag) corresponding to Inidividual tag

ScientificName

Value or column name in data set @ForestGeoSubset (Latin) corresponding to Scientific name

DateFormatMan

Value or column name in data set @ForestGeoSubset (yyyy-mm-dd) corresponding to What is your date format?

IdCensus

Value or column name in data set @ForestGeoSubset (CensusID) corresponding to Census ID

PlotAreaMan

Value or column name in data set @ForestGeoSubset (-999) corresponding to What is your plot area in hectare?

Year

Value or column name in data set @ForestGeoSubset (none) corresponding to Year of measurement

LifeStatus

Value or column name in data set @ForestGeoSubset (DFstatus) corresponding to Life status

PlotArea

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot area

CommercialSp

Value or column name in data set @ForestGeoSubset (none) corresponding to Commercial species indicator

SubplotAreaUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is the unit for Subplot area?

YTreeUTM

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree Y UTM coordinate

TreeUTMUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree UTM units

YearMan

Value or column name in data set @ForestGeoSubset (-999) corresponding to What is your year of measurement?

Circ

Value or column name in data set @ForestGeoSubset (none) corresponding to Circumference

SubplotAreaMan

Value or column name in data set @ForestGeoSubset (-999) corresponding to What is your Subplot area in hectare?

SubplotMan

Value or column name in data set @ForestGeoSubset () corresponding to What is your Subplot name ?

Site

Value or column name in data set @ForestGeoSubset (none) corresponding to Site name

TreeLon

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree longitude

Plot

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot name

IdTree

Value or column name in data set @ForestGeoSubset (treeID) corresponding to Individual ID

LifeForm

Value or column name in data set @ForestGeoSubset (none) corresponding to Life form

VernName

Value or column name in data set @ForestGeoSubset (none) corresponding to Vernacular name

SubplotArea

Value or column name in data set @ForestGeoSubset (none) corresponding to Subplot area

SiteMan

Value or column name in data set @ForestGeoSubset () corresponding to What is your site name ?

IsLiveMan

Value or column name in data set @ForestGeoSubset (c("alive", "broken below")) corresponding to Which of your status(es) represent a LIVE tree?

IsDeadMan

Value or column name in data set @ForestGeoSubset (stem dead) corresponding to Which of your status(es) represent a DEAD tree?

TreeCodes

Value or column name in data set @ForestGeoSubset (codes) corresponding to Select all the columns that give information about your tree, and/or your methods of measurements

TreeCodesSepMan

Value or column name in data set @ForestGeoSubset ([[:punct:]]) corresponding to If you store multiple tree codes in one cell, what character separates each code?

Authority

Value or column name in data set @ForestGeoSubset (SpeciesID) corresponding to Taxonomic authority for the ID level

TreeLat

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree latitude

Genus

Value or column name in data set @ForestGeoSubset (Genus) corresponding to Genus name

Subplot

Value or column name in data set @ForestGeoSubset (quadrat) corresponding to Subplot name

DiameterUnitMan

Value or column name in data set @ForestGeoSubset (mm) corresponding to What is your DBH unit ?

MinDBH

Value or column name in data set @ForestGeoSubset (none) corresponding to Minimum DBH

MinDBHMan

Value or column name in data set @ForestGeoSubset (10) corresponding to What is the minimum diameter of trees included in your inventory?

MinDBHUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your unit for the column indicating the minimum DBH? If not provided, it will be calculated looking at your Diameter column

XTreePlot

Value or column name in data set @ForestGeoSubset (gx) corresponding to Tree X euclidean position in plot

XTreeSubplot

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree X euclidean position in SUBplot

TreePlotUnitMan

Value or column name in data set @ForestGeoSubset (m) corresponding to Tree XY plot coordinates units

TreeSubplotUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree XY subplot coordinates units

Diameter

Value or column name in data set @ForestGeoSubset (dbh) corresponding to DBH

HOMUnitMan

Value or column name in data set @ForestGeoSubset (m) corresponding to What is your HOM unit?

Variety

Value or column name in data set @ForestGeoSubset (none) corresponding to Variety name

Voucher

Value or column name in data set @ForestGeoSubset (none) corresponding to Voucher code

SpeciesDensity

Value or column name in data set @ForestGeoSubset (none) corresponding to Species density

SpeciesBasalAreaDensity

Value or column name in data set @ForestGeoSubset (none) corresponding to Species basal area density

SpeciesBasalAreaDensityUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your species basal area unit?

SpeciesTotalBasalArea

Value or column name in data set @ForestGeoSubset (none) corresponding to Species total basal area

SpeciesTotalBasalAreaUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your basal area unit?

SpeciesVolume

Value or column name in data set @ForestGeoSubset (none) corresponding to Species volume

SpeciesVolumeUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your species volume unit?

PlotVolume

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot volume

PlotDensity

Value or column name in data set @ForestGeoSubset (none) corresponding to Tree Density

YPlotUTM

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot Y UTM coordinate

PlotElevation

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot elevation

SpeciesDensityUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your species density unit?

Subspecies

Value or column name in data set @ForestGeoSubset (none) corresponding to Subspecies name

PlotVolumeUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your plot volume unit?

PlotNDead

Value or column name in data set @ForestGeoSubset (none) corresponding to Number of dead trees in the sampling unit

PlotNLive

Value or column name in data set @ForestGeoSubset (none) corresponding to Number of living trees in the sampling unit

PlotSpeciesRichness

Value or column name in data set @ForestGeoSubset (none) corresponding to Number of species

XSubplotUTM

Value or column name in data set @ForestGeoSubset (none) corresponding to Subplot X UTM coordinate

PlotBasalAreaDensity

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot basal area density

YSubplotUTM

Value or column name in data set @ForestGeoSubset (none) corresponding to Subplot Y UTM coordinate

SpeciesCoverageIndex

Value or column name in data set @ForestGeoSubset (none) corresponding to Species Coverage Frequency

PlotElevationMan

Value or column name in data set @ForestGeoSubset (265) corresponding to What is your plot elevation in meters above sea level?

SpeciesRelativeDensity

Value or column name in data set @ForestGeoSubset (none) corresponding to Species Relative Density

XPlotUTM

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot X UTM coordinate

SubplotLon

Value or column name in data set @ForestGeoSubset (none) corresponding to Subplot longitude

PlotLon

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot longitude

SpeciesRelativeDominance

Value or column name in data set @ForestGeoSubset (none) corresponding to Species Relative Dominance

PlotDensityUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your plot density unit?

PlotBasalAreaDensityUnitMan

Value or column name in data set @ForestGeoSubset (none) corresponding to What is your plot basal area unit?

SpeciesImportanceIndex

Value or column name in data set @ForestGeoSubset (none) corresponding to Species Importance Index

PlotLat

Value or column name in data set @ForestGeoSubset (none) corresponding to Plot latitude

StemFieldNum

Value or column name in data set @ForestGeoSubset (StemTag) corresponding to Stem tag

SpeciesAbundance

Value or column name in data set @ForestGeoSubset (none) corresponding to Species abundance

SubplotLat

Value or column name in data set @ForestGeoSubset (none) corresponding to Subplot latitude

SpeciesRelativeFrequency

Value or column name in data set @ForestGeoSubset (none) corresponding to Species Relative Frequency

AllCodes

Some value created via interaction with the Shiny app

MeasLevel

Some value created via interaction with the Shiny app

Tidy

Some value created via interaction with the Shiny app

ClearValueName

Some value created via interaction with the Shiny app

TickedMelt3

Some value created via interaction with the Shiny app

Variablecolumns1

Some value created via interaction with the Shiny app

ValueName4

Some value created via interaction with the Shiny app

TickedMelt1

Some value created via interaction with the Shiny app

ValueName6

Some value created via interaction with the Shiny app

ValueName5

Some value created via interaction with the Shiny app

ValueName2

Some value created via interaction with the Shiny app

TickedMelt4

Some value created via interaction with the Shiny app

TickedMelt6

Some value created via interaction with the Shiny app

TickedMelt5

Some value created via interaction with the Shiny app

Variablecolumns3

Some value created via interaction with the Shiny app

ValueName3

Some value created via interaction with the Shiny app

TickedMelt2

Some value created via interaction with the Shiny app

Variablecolumns2

Some value created via interaction with the Shiny app

Variablecolumns4

Some value created via interaction with the Shiny app

ValueName1

Some value created via interaction with the Shiny app

...