ParacouProfile

Generated by interacting with app or function RequiredFormat_interactive

Usage

ParacouProfile

Format

A named list

ProfileName

Some value created via interaction with the Shiny app

TreeHeight

Value or column name in data set @ParacouSubset (none) corresponding to Tree height

XTreeUTM

Value or column name in data set @ParacouSubset (none) corresponding to Tree X UTM coordinate

Date

Value or column name in data set @ParacouSubset (CensusDate) corresponding to Date of measurement

SubplotAreaMan

Value or column name in data set @ParacouSubset (-999) corresponding to What is your Subplot area in hectare?

SpeciesVolume

Value or column name in data set @ParacouSubset (none) corresponding to Species volume

POM

Value or column name in data set @ParacouSubset (none) corresponding to POM

PlotAreaUnitMan

Value or column name in data set @ParacouSubset (ha) corresponding to What is the unit for plot area?

BHOMUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your BasalHOM unit?

PlotVolume

Value or column name in data set @ParacouSubset (none) corresponding to Plot volume

BD

Value or column name in data set @ParacouSubset (none) corresponding to BD

PlotDensity

Value or column name in data set @ParacouSubset (none) corresponding to Tree Density

IdStem

Value or column name in data set @ParacouSubset (none) corresponding to Stem ID

YTreeSubplot

Value or column name in data set @ParacouSubset (none) corresponding to Tree Y euclidean position in SUBplot

PlotMan

Value or column name in data set @ParacouSubset () corresponding to What is your plot name ?

YPlotUTM

Value or column name in data set @ParacouSubset (none) corresponding to Plot Y UTM coordinate

IdMeasure

Value or column name in data set @ParacouSubset (none) corresponding to Measure ID

ScientificName

Value or column name in data set @ParacouSubset (none) corresponding to Scientific name

SpeciesBasalAreaDensityUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your species basal area unit?

SubplotArea

Value or column name in data set @ParacouSubset (none) corresponding to Subplot area

IdCensus

Value or column name in data set @ParacouSubset (none) corresponding to Census ID

XTreePlot

Value or column name in data set @ParacouSubset (none) corresponding to Tree X euclidean position in plot

PlotAreaMan

Value or column name in data set @ParacouSubset (-999) corresponding to What is your plot area in hectare?

PlotElevation

Value or column name in data set @ParacouSubset (none) corresponding to Plot elevation

BPOM

Value or column name in data set @ParacouSubset (none) corresponding to Basal POM

SpeciesDensityUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your species density unit?

CommercialSp

Value or column name in data set @ParacouSubset (CommercialSp) corresponding to Commercial species indicator

IsCommercialMan

Value or column name in data set @ParacouSubset (TRUE) corresponding to Which of the following(s) indicate a commercial species?

SpeciesBasalAreaDensity

Value or column name in data set @ParacouSubset (none) corresponding to Species basal area density

YearMan

Value or column name in data set @ParacouSubset (-999) corresponding to What is your year of measurement?

Subspecies

Value or column name in data set @ParacouSubset (none) corresponding to Subspecies name

Circ

Value or column name in data set @ParacouSubset (Circ) corresponding to Circumference

SpeciesTotalBasalAreaUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your basal area unit?

PlotVolumeUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your plot volume unit?

PlotNDead

Value or column name in data set @ParacouSubset (none) corresponding to Number of dead trees in the sampling unit

XTreeSubplot

Value or column name in data set @ParacouSubset (none) corresponding to Tree X euclidean position in SUBplot

Site

Value or column name in data set @ParacouSubset (Forest) corresponding to Site name

Cluster

Value or column name in data set @ParacouSubset (none) corresponding to Cluster of plots

MinDBH

Value or column name in data set @ParacouSubset (none) corresponding to Minimum DBH

Day

Value or column name in data set @ParacouSubset (none) corresponding to Day of measurement

TreeUTMUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to Tree UTM units

PlotNLive

Value or column name in data set @ParacouSubset (none) corresponding to Number of living trees in the sampling unit

SpeciesVolumeUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your species volume unit?

BHOM

Value or column name in data set @ParacouSubset (none) corresponding to Basal HOM

IdTree

Value or column name in data set @ParacouSubset (idTree) corresponding to Individual ID

LifeForm

Value or column name in data set @ParacouSubset (none) corresponding to Life form

TreeSubplotUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to Tree XY subplot coordinates units

VernName

Value or column name in data set @ParacouSubset (VernName) corresponding to Vernacular name

PlotSpeciesRichness

Value or column name in data set @ParacouSubset (none) corresponding to Number of species

SiteMan

Value or column name in data set @ParacouSubset () corresponding to What is your site name ?

XSubplotUTM

Value or column name in data set @ParacouSubset (none) corresponding to Subplot X UTM coordinate

Authority

Value or column name in data set @ParacouSubset (none) corresponding to Taxonomic authority for the ID level

BCirc

Value or column name in data set @ParacouSubset (none) corresponding to Basal Circumference

PlotBasalAreaDensity

Value or column name in data set @ParacouSubset (none) corresponding to Plot basal area density

TreePlotUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to Tree XY plot coordinates units

Diameter

Value or column name in data set @ParacouSubset (none) corresponding to DBH

Month

Value or column name in data set @ParacouSubset (none) corresponding to Month of measurement

HOMUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your HOM unit?

SpeciesTotalBasalArea

Value or column name in data set @ParacouSubset (none) corresponding to Species total basal area

TreeHeightUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your tree height unit?

Variety

Value or column name in data set @ParacouSubset (none) corresponding to Variety name

Family

Value or column name in data set @ParacouSubset (Family) corresponding to Family name

HOM

Value or column name in data set @ParacouSubset (none) corresponding to HOM

HOMMan

Value or column name in data set @ParacouSubset (1.3) corresponding to What is the height of diameter measurement in your inventory?

ScientificNameSepMan

Value or column name in data set @ParacouSubset () corresponding to Character separating Genus and Species in Scientific name

BDUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your BD unit ?

YSubplotUTM

Value or column name in data set @ParacouSubset (none) corresponding to Subplot Y UTM coordinate

SpeciesCoverageIndex

Value or column name in data set @ParacouSubset (none) corresponding to Species Coverage Frequency

DiameterUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your DBH unit ?

PlotElevationMan

Value or column name in data set @ParacouSubset (-999) corresponding to What is your plot elevation in meters above sea level?

SubplotMan

Value or column name in data set @ParacouSubset () corresponding to What is your Subplot name ?

DateFormatMan

Value or column name in data set @ParacouSubset (yyyy-mm-dd) corresponding to What is your date format?

Species

Value or column name in data set @ParacouSubset (Species) corresponding to Species name

SpeciesRelativeDensity

Value or column name in data set @ParacouSubset (none) corresponding to Species Relative Density

CircUnitMan

Value or column name in data set @ParacouSubset (cm) corresponding to What is your circumference unit?

IdLevel

Value or column name in data set @ParacouSubset (none) corresponding to Determination rank

XPlotUTM

Value or column name in data set @ParacouSubset (none) corresponding to Plot X UTM coordinate

TreeFieldNum

Value or column name in data set @ParacouSubset (TreeFieldNum) corresponding to Inidividual tag

YTreeUTM

Value or column name in data set @ParacouSubset (none) corresponding to Tree Y UTM coordinate

TreeLon

Value or column name in data set @ParacouSubset (none) corresponding to Tree longitude

BCircUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your basal circumference unit?

SubplotLon

Value or column name in data set @ParacouSubset (none) corresponding to Subplot longitude

Year

Value or column name in data set @ParacouSubset (CensusYear) corresponding to Year of measurement

LifeStatus

Value or column name in data set @ParacouSubset (CodeAlive) corresponding to Live status

DeadStatus

Value or column name in data set @ParacouSubset (CodeAlive) corresponding to Dead status

MinDBHMan

Value or column name in data set @ParacouSubset (-999) corresponding to What is the minimum diameter of trees included in your inventory?

PlotArea

Value or column name in data set @ParacouSubset (PlotArea) corresponding to Plot area

PlotLon

Value or column name in data set @ParacouSubset (none) corresponding to Plot longitude

YTreePlot

Value or column name in data set @ParacouSubset (none) corresponding to Tree Y euclidean position in plot

SpeciesRelativeDominance

Value or column name in data set @ParacouSubset (none) corresponding to Species Relative Dominance

Subplot

Value or column name in data set @ParacouSubset (SubPlot) corresponding to Subplot name

MinDBHUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your unit for the column indicating the minimum DBH? If not provided, it will be calculated looking at your Diameter column

PlotDensityUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your plot density unit?

PlotBasalAreaDensityUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is your plot basal area unit?

Plot

Value or column name in data set @ParacouSubset (Plot) corresponding to Plot name

SpeciesImportanceIndex

Value or column name in data set @ParacouSubset (none) corresponding to Species Importance Index

SubplotAreaUnitMan

Value or column name in data set @ParacouSubset (none) corresponding to What is the unit for Subplot area?

Voucher

Value or column name in data set @ParacouSubset (none) corresponding to Voucher code

TreeLat

Value or column name in data set @ParacouSubset (none) corresponding to Tree latitude

PlotLat

Value or column name in data set @ParacouSubset (none) corresponding to Plot latitude

StemFieldNum

Value or column name in data set @ParacouSubset (none) corresponding to Stem tag

Genus

Value or column name in data set @ParacouSubset (Genus) corresponding to Genus name

SpeciesAbundance

Value or column name in data set @ParacouSubset (none) corresponding to Species abundance

IsLiveMan

Value or column name in data set @ParacouSubset (TRUE) corresponding to Which of your status(es) represent a LIVE tree?

IsDeadMan

Value or column name in data set @ParacouSubset (FALSE) corresponding to Which of your status(es) represent a DEAD tree?

SubplotLat

Value or column name in data set @ParacouSubset (none) corresponding to Subplot latitude

SpeciesDensity

Value or column name in data set @ParacouSubset (none) corresponding to Species density

SpeciesRelativeFrequency

Value or column name in data set @ParacouSubset (none) corresponding to Species Relative Frequency

MeasLevel

Some value created via interaction with the Shiny app

Tidy

Some value created via interaction with the Shiny app

ClearValueName

Some value created via interaction with the Shiny app

TickedMelt3

Some value created via interaction with the Shiny app

Variablecolumns1

Some value created via interaction with the Shiny app

ValueName4

Some value created via interaction with the Shiny app

TickedMelt1

Some value created via interaction with the Shiny app

ValueName5

Some value created via interaction with the Shiny app

ValueName2

Some value created via interaction with the Shiny app

TickedMelt4

Some value created via interaction with the Shiny app

TickedMelt5

Some value created via interaction with the Shiny app

Variablecolumns3

Some value created via interaction with the Shiny app

ValueName3

Some value created via interaction with the Shiny app

TickedMelt2

Some value created via interaction with the Shiny app

Variablecolumns2

Some value created via interaction with the Shiny app

ValueName1

Some value created via interaction with the Shiny app

AllCodes

Some value created via interaction with the Shiny app

...