Status Correction

Detect errors, or detect errors and correct, the tree life status evolution over the censuses. Inspired by the code of Nino Page package (ForestData::correct_alive() and .correct_alive_tree())

Usage

StatusCorrection(
  Data,
  DeathConfirmation = 2,
  UseSize = FALSE,
  AddRowsForForgottenCensuses = TRUE,
  KeepMeas = c("MaxHOM", "MaxDate")
)

Arguments

Data

Dataset (data.frame or data.table) The LifeStatus column must be coded as: - TRUE = alive, - FALSE = dead, - NA = unseen

DeathConfirmation

Number of times (censuses) needed for an unseen tree to be considered dead (numeric) (Default = 2 censuses)

UseSize

Use the size presence (> min DBH) as a witness of the living status of the tree (logical) (Default = FALSE)

AddRowsForForgottenCensuses

TRUE: adds rows for forgotten censuses, FALSE: does not add any rows (logical).

KeepMeas

In case of multiple diameter measurements in the same census year: Possible values: "MaxHOM", "MaxDate" (character).

"MaxHOM": apply the correction to the measurement taken at the highest POM

"MaxDate": apply the correction to the most recent measurement (same year but more recent date)

Value

Fill the Comment_DataHarmonization column with error type information and add a LifeStatus_DataHarmonizationCor column with corrected trees life status. If AddRowsForForgottenCensuses is TRUE, a column MissedStem_DataHarmonizationCor is added with value TRUE for new rows.

Details

if UseSize : if Diameter != NA -> Alive If (the value in bold is modified by the value given after the arrow): (the ">" gives the chronological order of the sequence)

Dead > Alive -> NA

add rows for the forgotten censuses between 2 'Alive' if chosen

Alive > Dead/NA > Alive -> Alive

Alive > NA > Dead -> NA

Alive > Dead > NA -> Dead

Alive > NA > NA: if DeathConfirmation > unseens -> NA if DeathConfirmation =< unseens -> Dead

Examples

library(data.table)
data(TestData)

selection <- c("101184", "101433","101435","101436")

# Write the sequence
TestData <- TestData[order(IdCensus)] # arrange IdCensus in ascending order
TestData[IdTree == "101184", LifeStatus := c(TRUE, TRUE, TRUE, TRUE, FALSE)]
#>       MinDBH IdCensus Year Month Day       Date    Site Cluster Plot PlotViewID
#>    1:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    2:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    3:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    4:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    5:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>   ---                                                                          
#> 4851:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4852:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4853:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4854:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4855:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#>       Subplot PlotArea PlotElevation SubplotArea PlotLat PlotLon XPlotUTM
#>    1:       1     6.25            NA           0      NA      NA       NA
#>    2:       1     6.25            NA           0      NA      NA       NA
#>    3:       1     6.25            NA           0      NA      NA       NA
#>    4:       1     6.25            NA           0      NA      NA       NA
#>    5:       1     6.25            NA           0      NA      NA       NA
#>   ---                                                                    
#> 4851:       1     6.25            NA           0      NA      NA       NA
#> 4852:       1     6.25            NA           0      NA      NA       NA
#> 4853:       1     6.25            NA           0      NA      NA       NA
#> 4854:       1     6.25            NA           0      NA      NA       NA
#> 4855:       1     6.25            NA           0      NA      NA       NA
#>       YPlotUTM SubplotLat SubplotLon XSubplotUTM YSubplotUTM
#>    1:       NA         NA         NA          NA          NA
#>    2:       NA         NA         NA          NA          NA
#>    3:       NA         NA         NA          NA          NA
#>    4:       NA         NA         NA          NA          NA
#>    5:       NA         NA         NA          NA          NA
#>   ---                                                       
#> 4851:       NA         NA         NA          NA          NA
#> 4852:       NA         NA         NA          NA          NA
#> 4853:       NA         NA         NA          NA          NA
#> 4854:       NA         NA         NA          NA          NA
#> 4855:       NA         NA         NA          NA          NA
#>               ScientificName   VernName           Family       Genus    Species
#>    1:         Eperua falcata       wapa         Fabaceae      Eperua    falcata
#>    2:  Conceveiba guianensis     kusisi    Euphorbiaceae  Conceveiba guianensis
#>    3:     Moronobea coccinea  Moronobea       Clusiaceae   Moronobea   coccinea
#>    4:    Pouteria melanopoda    bakuman       Sapotaceae    Pouteria melanopoda
#>    5:  Sandwithia guyanensis  wata tiki    Euphorbiaceae  Sandwithia guyanensis
#>   ---                                                                          
#> 4851: Micropholis guyanensis    bakuman       Sapotaceae Micropholis guyanensis
#> 4852:    Lecythis persistens maho rouge    Lecythidaceae    Lecythis persistens
#> 4853:    Lecythis persistens maho rouge    Lecythidaceae    Lecythis persistens
#> 4854:         Eperua falcata       wapa         Fabaceae      Eperua    falcata
#> 4855:           Licania alba       koko Chrysobalanaceae     Licania       alba
#>       Subspecies Variety Voucher IdLevel Authority CommercialSp LifeForm
#>    1:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#>    2:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>    3:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#>    4:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>    5:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>   ---                                                                   
#> 4851:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4852:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4853:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4854:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#> 4855:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>       TreeFieldNum IdTree StemFieldNum        IdStem TreeLat TreeLon XTreeUTM
#>    1:           45 100665         <NA> 100665_1_auto      NA      NA       NA
#>    2:          148 100765         <NA> 100765_1_auto      NA      NA       NA
#>    3:           46 100666         <NA> 100666_1_auto      NA      NA       NA
#>    4:           49 100669         <NA> 100669_1_auto      NA      NA       NA
#>    5:            1 100621         <NA> 100621_1_auto      NA      NA       NA
#>   ---                                                                        
#> 4851:          922 101517         <NA> 101517_1_auto      NA      NA       NA
#> 4852:          884 101479         <NA> 101479_1_auto      NA      NA       NA
#> 4853:          887 101482         <NA> 101482_1_auto      NA      NA       NA
#> 4854:          896 101491         <NA> 101491_1_auto      NA      NA       NA
#> 4855:          905 101500         <NA> 101500_1_auto      NA      NA       NA
#>       YTreeUTM XTreePlot YTreePlot XTreeSubplot YTreeSubplot LifeStatus
#>    1:       NA        NA        NA           NA           NA       TRUE
#>    2:       NA        NA        NA           NA           NA       TRUE
#>    3:       NA        NA        NA           NA           NA       TRUE
#>    4:       NA        NA        NA           NA           NA       TRUE
#>    5:       NA        NA        NA           NA           NA       TRUE
#>   ---                                                                  
#> 4851:       NA        NA        NA           NA           NA       TRUE
#> 4852:       NA        NA        NA           NA           NA       TRUE
#> 4853:       NA        NA        NA           NA           NA       TRUE
#> 4854:       NA        NA        NA           NA           NA       TRUE
#> 4855:       NA        NA        NA           NA           NA       TRUE
#>       Diameter BD  Circ BCirc HOM  POM BHOM BPOM TreeHeight
#>    1:    53.16 NA 167.0    NA 1.3 <NA>   NA <NA>         NA
#>    2:    14.32 NA  45.0    NA 1.3 <NA>   NA <NA>         NA
#>    3:    19.89 NA  62.5    NA 1.3 <NA>   NA <NA>         NA
#>    4:    41.38 NA 130.0    NA 1.3 <NA>   NA <NA>         NA
#>    5:    13.37 NA  42.0    NA 1.3 <NA>   NA <NA>         NA
#>   ---                                                      
#> 4851:    22.60 NA  71.0    NA 1.3 <NA>   NA <NA>         NA
#> 4852:    16.39 NA  51.5    NA 1.3 <NA>   NA <NA>         NA
#> 4853:    20.85 NA  65.5    NA 1.3 <NA>   NA <NA>         NA
#> 4854:    29.60 NA  93.0    NA 1.3 <NA>   NA <NA>         NA
#> 4855:    27.53 NA  86.5    NA 1.3 <NA>   NA <NA>         NA
#>       TreeFieldNumOriginal IdTreeOriginal CommercialSpOriginal DateOriginal
#>    1:                   45         100665                 TRUE   2016-09-14
#>    2:                  148         100765                FALSE   2016-09-14
#>    3:                   46         100666                 TRUE   2016-09-14
#>    4:                   49         100669                FALSE   2016-09-14
#>    5:                    1         100621                FALSE   2016-09-14
#>   ---                                                                      
#> 4851:                  922         101517                FALSE   2020-06-04
#> 4852:                  884         101479                FALSE   2020-06-04
#> 4853:                  887         101482                FALSE   2020-06-04
#> 4854:                  896         101491                 TRUE   2020-06-04
#> 4855:                  905         101500                FALSE   2020-06-04
#>       LifeStatusOriginal
#>    1:               TRUE
#>    2:               TRUE
#>    3:               TRUE
#>    4:               TRUE
#>    5:               TRUE
#>   ---                   
#> 4851:               TRUE
#> 4852:               TRUE
#> 4853:               TRUE
#> 4854:               TRUE
#> 4855:               TRUE
TestData[IdTree == "101433", LifeStatus := c(FALSE, TRUE, TRUE, TRUE, TRUE)]
#>       MinDBH IdCensus Year Month Day       Date    Site Cluster Plot PlotViewID
#>    1:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    2:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    3:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    4:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    5:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>   ---                                                                          
#> 4851:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4852:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4853:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4854:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4855:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#>       Subplot PlotArea PlotElevation SubplotArea PlotLat PlotLon XPlotUTM
#>    1:       1     6.25            NA           0      NA      NA       NA
#>    2:       1     6.25            NA           0      NA      NA       NA
#>    3:       1     6.25            NA           0      NA      NA       NA
#>    4:       1     6.25            NA           0      NA      NA       NA
#>    5:       1     6.25            NA           0      NA      NA       NA
#>   ---                                                                    
#> 4851:       1     6.25            NA           0      NA      NA       NA
#> 4852:       1     6.25            NA           0      NA      NA       NA
#> 4853:       1     6.25            NA           0      NA      NA       NA
#> 4854:       1     6.25            NA           0      NA      NA       NA
#> 4855:       1     6.25            NA           0      NA      NA       NA
#>       YPlotUTM SubplotLat SubplotLon XSubplotUTM YSubplotUTM
#>    1:       NA         NA         NA          NA          NA
#>    2:       NA         NA         NA          NA          NA
#>    3:       NA         NA         NA          NA          NA
#>    4:       NA         NA         NA          NA          NA
#>    5:       NA         NA         NA          NA          NA
#>   ---                                                       
#> 4851:       NA         NA         NA          NA          NA
#> 4852:       NA         NA         NA          NA          NA
#> 4853:       NA         NA         NA          NA          NA
#> 4854:       NA         NA         NA          NA          NA
#> 4855:       NA         NA         NA          NA          NA
#>               ScientificName   VernName           Family       Genus    Species
#>    1:         Eperua falcata       wapa         Fabaceae      Eperua    falcata
#>    2:  Conceveiba guianensis     kusisi    Euphorbiaceae  Conceveiba guianensis
#>    3:     Moronobea coccinea  Moronobea       Clusiaceae   Moronobea   coccinea
#>    4:    Pouteria melanopoda    bakuman       Sapotaceae    Pouteria melanopoda
#>    5:  Sandwithia guyanensis  wata tiki    Euphorbiaceae  Sandwithia guyanensis
#>   ---                                                                          
#> 4851: Micropholis guyanensis    bakuman       Sapotaceae Micropholis guyanensis
#> 4852:    Lecythis persistens maho rouge    Lecythidaceae    Lecythis persistens
#> 4853:    Lecythis persistens maho rouge    Lecythidaceae    Lecythis persistens
#> 4854:         Eperua falcata       wapa         Fabaceae      Eperua    falcata
#> 4855:           Licania alba       koko Chrysobalanaceae     Licania       alba
#>       Subspecies Variety Voucher IdLevel Authority CommercialSp LifeForm
#>    1:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#>    2:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>    3:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#>    4:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>    5:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>   ---                                                                   
#> 4851:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4852:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4853:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4854:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#> 4855:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>       TreeFieldNum IdTree StemFieldNum        IdStem TreeLat TreeLon XTreeUTM
#>    1:           45 100665         <NA> 100665_1_auto      NA      NA       NA
#>    2:          148 100765         <NA> 100765_1_auto      NA      NA       NA
#>    3:           46 100666         <NA> 100666_1_auto      NA      NA       NA
#>    4:           49 100669         <NA> 100669_1_auto      NA      NA       NA
#>    5:            1 100621         <NA> 100621_1_auto      NA      NA       NA
#>   ---                                                                        
#> 4851:          922 101517         <NA> 101517_1_auto      NA      NA       NA
#> 4852:          884 101479         <NA> 101479_1_auto      NA      NA       NA
#> 4853:          887 101482         <NA> 101482_1_auto      NA      NA       NA
#> 4854:          896 101491         <NA> 101491_1_auto      NA      NA       NA
#> 4855:          905 101500         <NA> 101500_1_auto      NA      NA       NA
#>       YTreeUTM XTreePlot YTreePlot XTreeSubplot YTreeSubplot LifeStatus
#>    1:       NA        NA        NA           NA           NA       TRUE
#>    2:       NA        NA        NA           NA           NA       TRUE
#>    3:       NA        NA        NA           NA           NA       TRUE
#>    4:       NA        NA        NA           NA           NA       TRUE
#>    5:       NA        NA        NA           NA           NA       TRUE
#>   ---                                                                  
#> 4851:       NA        NA        NA           NA           NA       TRUE
#> 4852:       NA        NA        NA           NA           NA       TRUE
#> 4853:       NA        NA        NA           NA           NA       TRUE
#> 4854:       NA        NA        NA           NA           NA       TRUE
#> 4855:       NA        NA        NA           NA           NA       TRUE
#>       Diameter BD  Circ BCirc HOM  POM BHOM BPOM TreeHeight
#>    1:    53.16 NA 167.0    NA 1.3 <NA>   NA <NA>         NA
#>    2:    14.32 NA  45.0    NA 1.3 <NA>   NA <NA>         NA
#>    3:    19.89 NA  62.5    NA 1.3 <NA>   NA <NA>         NA
#>    4:    41.38 NA 130.0    NA 1.3 <NA>   NA <NA>         NA
#>    5:    13.37 NA  42.0    NA 1.3 <NA>   NA <NA>         NA
#>   ---                                                      
#> 4851:    22.60 NA  71.0    NA 1.3 <NA>   NA <NA>         NA
#> 4852:    16.39 NA  51.5    NA 1.3 <NA>   NA <NA>         NA
#> 4853:    20.85 NA  65.5    NA 1.3 <NA>   NA <NA>         NA
#> 4854:    29.60 NA  93.0    NA 1.3 <NA>   NA <NA>         NA
#> 4855:    27.53 NA  86.5    NA 1.3 <NA>   NA <NA>         NA
#>       TreeFieldNumOriginal IdTreeOriginal CommercialSpOriginal DateOriginal
#>    1:                   45         100665                 TRUE   2016-09-14
#>    2:                  148         100765                FALSE   2016-09-14
#>    3:                   46         100666                 TRUE   2016-09-14
#>    4:                   49         100669                FALSE   2016-09-14
#>    5:                    1         100621                FALSE   2016-09-14
#>   ---                                                                      
#> 4851:                  922         101517                FALSE   2020-06-04
#> 4852:                  884         101479                FALSE   2020-06-04
#> 4853:                  887         101482                FALSE   2020-06-04
#> 4854:                  896         101491                 TRUE   2020-06-04
#> 4855:                  905         101500                FALSE   2020-06-04
#>       LifeStatusOriginal
#>    1:               TRUE
#>    2:               TRUE
#>    3:               TRUE
#>    4:               TRUE
#>    5:               TRUE
#>   ---                   
#> 4851:               TRUE
#> 4852:               TRUE
#> 4853:               TRUE
#> 4854:               TRUE
#> 4855:               TRUE
TestData[IdTree == "101435", LifeStatus := c(TRUE, TRUE, NA, FALSE, TRUE)]
#>       MinDBH IdCensus Year Month Day       Date    Site Cluster Plot PlotViewID
#>    1:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    2:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    3:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    4:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    5:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>   ---                                                                          
#> 4851:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4852:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4853:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4854:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4855:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#>       Subplot PlotArea PlotElevation SubplotArea PlotLat PlotLon XPlotUTM
#>    1:       1     6.25            NA           0      NA      NA       NA
#>    2:       1     6.25            NA           0      NA      NA       NA
#>    3:       1     6.25            NA           0      NA      NA       NA
#>    4:       1     6.25            NA           0      NA      NA       NA
#>    5:       1     6.25            NA           0      NA      NA       NA
#>   ---                                                                    
#> 4851:       1     6.25            NA           0      NA      NA       NA
#> 4852:       1     6.25            NA           0      NA      NA       NA
#> 4853:       1     6.25            NA           0      NA      NA       NA
#> 4854:       1     6.25            NA           0      NA      NA       NA
#> 4855:       1     6.25            NA           0      NA      NA       NA
#>       YPlotUTM SubplotLat SubplotLon XSubplotUTM YSubplotUTM
#>    1:       NA         NA         NA          NA          NA
#>    2:       NA         NA         NA          NA          NA
#>    3:       NA         NA         NA          NA          NA
#>    4:       NA         NA         NA          NA          NA
#>    5:       NA         NA         NA          NA          NA
#>   ---                                                       
#> 4851:       NA         NA         NA          NA          NA
#> 4852:       NA         NA         NA          NA          NA
#> 4853:       NA         NA         NA          NA          NA
#> 4854:       NA         NA         NA          NA          NA
#> 4855:       NA         NA         NA          NA          NA
#>               ScientificName   VernName           Family       Genus    Species
#>    1:         Eperua falcata       wapa         Fabaceae      Eperua    falcata
#>    2:  Conceveiba guianensis     kusisi    Euphorbiaceae  Conceveiba guianensis
#>    3:     Moronobea coccinea  Moronobea       Clusiaceae   Moronobea   coccinea
#>    4:    Pouteria melanopoda    bakuman       Sapotaceae    Pouteria melanopoda
#>    5:  Sandwithia guyanensis  wata tiki    Euphorbiaceae  Sandwithia guyanensis
#>   ---                                                                          
#> 4851: Micropholis guyanensis    bakuman       Sapotaceae Micropholis guyanensis
#> 4852:    Lecythis persistens maho rouge    Lecythidaceae    Lecythis persistens
#> 4853:    Lecythis persistens maho rouge    Lecythidaceae    Lecythis persistens
#> 4854:         Eperua falcata       wapa         Fabaceae      Eperua    falcata
#> 4855:           Licania alba       koko Chrysobalanaceae     Licania       alba
#>       Subspecies Variety Voucher IdLevel Authority CommercialSp LifeForm
#>    1:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#>    2:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>    3:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#>    4:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>    5:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>   ---                                                                   
#> 4851:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4852:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4853:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4854:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#> 4855:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>       TreeFieldNum IdTree StemFieldNum        IdStem TreeLat TreeLon XTreeUTM
#>    1:           45 100665         <NA> 100665_1_auto      NA      NA       NA
#>    2:          148 100765         <NA> 100765_1_auto      NA      NA       NA
#>    3:           46 100666         <NA> 100666_1_auto      NA      NA       NA
#>    4:           49 100669         <NA> 100669_1_auto      NA      NA       NA
#>    5:            1 100621         <NA> 100621_1_auto      NA      NA       NA
#>   ---                                                                        
#> 4851:          922 101517         <NA> 101517_1_auto      NA      NA       NA
#> 4852:          884 101479         <NA> 101479_1_auto      NA      NA       NA
#> 4853:          887 101482         <NA> 101482_1_auto      NA      NA       NA
#> 4854:          896 101491         <NA> 101491_1_auto      NA      NA       NA
#> 4855:          905 101500         <NA> 101500_1_auto      NA      NA       NA
#>       YTreeUTM XTreePlot YTreePlot XTreeSubplot YTreeSubplot LifeStatus
#>    1:       NA        NA        NA           NA           NA       TRUE
#>    2:       NA        NA        NA           NA           NA       TRUE
#>    3:       NA        NA        NA           NA           NA       TRUE
#>    4:       NA        NA        NA           NA           NA       TRUE
#>    5:       NA        NA        NA           NA           NA       TRUE
#>   ---                                                                  
#> 4851:       NA        NA        NA           NA           NA       TRUE
#> 4852:       NA        NA        NA           NA           NA       TRUE
#> 4853:       NA        NA        NA           NA           NA       TRUE
#> 4854:       NA        NA        NA           NA           NA       TRUE
#> 4855:       NA        NA        NA           NA           NA       TRUE
#>       Diameter BD  Circ BCirc HOM  POM BHOM BPOM TreeHeight
#>    1:    53.16 NA 167.0    NA 1.3 <NA>   NA <NA>         NA
#>    2:    14.32 NA  45.0    NA 1.3 <NA>   NA <NA>         NA
#>    3:    19.89 NA  62.5    NA 1.3 <NA>   NA <NA>         NA
#>    4:    41.38 NA 130.0    NA 1.3 <NA>   NA <NA>         NA
#>    5:    13.37 NA  42.0    NA 1.3 <NA>   NA <NA>         NA
#>   ---                                                      
#> 4851:    22.60 NA  71.0    NA 1.3 <NA>   NA <NA>         NA
#> 4852:    16.39 NA  51.5    NA 1.3 <NA>   NA <NA>         NA
#> 4853:    20.85 NA  65.5    NA 1.3 <NA>   NA <NA>         NA
#> 4854:    29.60 NA  93.0    NA 1.3 <NA>   NA <NA>         NA
#> 4855:    27.53 NA  86.5    NA 1.3 <NA>   NA <NA>         NA
#>       TreeFieldNumOriginal IdTreeOriginal CommercialSpOriginal DateOriginal
#>    1:                   45         100665                 TRUE   2016-09-14
#>    2:                  148         100765                FALSE   2016-09-14
#>    3:                   46         100666                 TRUE   2016-09-14
#>    4:                   49         100669                FALSE   2016-09-14
#>    5:                    1         100621                FALSE   2016-09-14
#>   ---                                                                      
#> 4851:                  922         101517                FALSE   2020-06-04
#> 4852:                  884         101479                FALSE   2020-06-04
#> 4853:                  887         101482                FALSE   2020-06-04
#> 4854:                  896         101491                 TRUE   2020-06-04
#> 4855:                  905         101500                FALSE   2020-06-04
#>       LifeStatusOriginal
#>    1:               TRUE
#>    2:               TRUE
#>    3:               TRUE
#>    4:               TRUE
#>    5:               TRUE
#>   ---                   
#> 4851:               TRUE
#> 4852:               TRUE
#> 4853:               TRUE
#> 4854:               TRUE
#> 4855:               TRUE
TestData[IdTree == "101436", LifeStatus := c(TRUE, NA, NA, FALSE, NA)]
#>       MinDBH IdCensus Year Month Day       Date    Site Cluster Plot PlotViewID
#>    1:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    2:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    3:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    4:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    5:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>   ---                                                                          
#> 4851:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4852:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4853:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4854:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4855:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#>       Subplot PlotArea PlotElevation SubplotArea PlotLat PlotLon XPlotUTM
#>    1:       1     6.25            NA           0      NA      NA       NA
#>    2:       1     6.25            NA           0      NA      NA       NA
#>    3:       1     6.25            NA           0      NA      NA       NA
#>    4:       1     6.25            NA           0      NA      NA       NA
#>    5:       1     6.25            NA           0      NA      NA       NA
#>   ---                                                                    
#> 4851:       1     6.25            NA           0      NA      NA       NA
#> 4852:       1     6.25            NA           0      NA      NA       NA
#> 4853:       1     6.25            NA           0      NA      NA       NA
#> 4854:       1     6.25            NA           0      NA      NA       NA
#> 4855:       1     6.25            NA           0      NA      NA       NA
#>       YPlotUTM SubplotLat SubplotLon XSubplotUTM YSubplotUTM
#>    1:       NA         NA         NA          NA          NA
#>    2:       NA         NA         NA          NA          NA
#>    3:       NA         NA         NA          NA          NA
#>    4:       NA         NA         NA          NA          NA
#>    5:       NA         NA         NA          NA          NA
#>   ---                                                       
#> 4851:       NA         NA         NA          NA          NA
#> 4852:       NA         NA         NA          NA          NA
#> 4853:       NA         NA         NA          NA          NA
#> 4854:       NA         NA         NA          NA          NA
#> 4855:       NA         NA         NA          NA          NA
#>               ScientificName   VernName           Family       Genus    Species
#>    1:         Eperua falcata       wapa         Fabaceae      Eperua    falcata
#>    2:  Conceveiba guianensis     kusisi    Euphorbiaceae  Conceveiba guianensis
#>    3:     Moronobea coccinea  Moronobea       Clusiaceae   Moronobea   coccinea
#>    4:    Pouteria melanopoda    bakuman       Sapotaceae    Pouteria melanopoda
#>    5:  Sandwithia guyanensis  wata tiki    Euphorbiaceae  Sandwithia guyanensis
#>   ---                                                                          
#> 4851: Micropholis guyanensis    bakuman       Sapotaceae Micropholis guyanensis
#> 4852:    Lecythis persistens maho rouge    Lecythidaceae    Lecythis persistens
#> 4853:    Lecythis persistens maho rouge    Lecythidaceae    Lecythis persistens
#> 4854:         Eperua falcata       wapa         Fabaceae      Eperua    falcata
#> 4855:           Licania alba       koko Chrysobalanaceae     Licania       alba
#>       Subspecies Variety Voucher IdLevel Authority CommercialSp LifeForm
#>    1:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#>    2:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>    3:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#>    4:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>    5:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>   ---                                                                   
#> 4851:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4852:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4853:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#> 4854:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     tree
#> 4855:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     tree
#>       TreeFieldNum IdTree StemFieldNum        IdStem TreeLat TreeLon XTreeUTM
#>    1:           45 100665         <NA> 100665_1_auto      NA      NA       NA
#>    2:          148 100765         <NA> 100765_1_auto      NA      NA       NA
#>    3:           46 100666         <NA> 100666_1_auto      NA      NA       NA
#>    4:           49 100669         <NA> 100669_1_auto      NA      NA       NA
#>    5:            1 100621         <NA> 100621_1_auto      NA      NA       NA
#>   ---                                                                        
#> 4851:          922 101517         <NA> 101517_1_auto      NA      NA       NA
#> 4852:          884 101479         <NA> 101479_1_auto      NA      NA       NA
#> 4853:          887 101482         <NA> 101482_1_auto      NA      NA       NA
#> 4854:          896 101491         <NA> 101491_1_auto      NA      NA       NA
#> 4855:          905 101500         <NA> 101500_1_auto      NA      NA       NA
#>       YTreeUTM XTreePlot YTreePlot XTreeSubplot YTreeSubplot LifeStatus
#>    1:       NA        NA        NA           NA           NA       TRUE
#>    2:       NA        NA        NA           NA           NA       TRUE
#>    3:       NA        NA        NA           NA           NA       TRUE
#>    4:       NA        NA        NA           NA           NA       TRUE
#>    5:       NA        NA        NA           NA           NA       TRUE
#>   ---                                                                  
#> 4851:       NA        NA        NA           NA           NA       TRUE
#> 4852:       NA        NA        NA           NA           NA       TRUE
#> 4853:       NA        NA        NA           NA           NA       TRUE
#> 4854:       NA        NA        NA           NA           NA       TRUE
#> 4855:       NA        NA        NA           NA           NA       TRUE
#>       Diameter BD  Circ BCirc HOM  POM BHOM BPOM TreeHeight
#>    1:    53.16 NA 167.0    NA 1.3 <NA>   NA <NA>         NA
#>    2:    14.32 NA  45.0    NA 1.3 <NA>   NA <NA>         NA
#>    3:    19.89 NA  62.5    NA 1.3 <NA>   NA <NA>         NA
#>    4:    41.38 NA 130.0    NA 1.3 <NA>   NA <NA>         NA
#>    5:    13.37 NA  42.0    NA 1.3 <NA>   NA <NA>         NA
#>   ---                                                      
#> 4851:    22.60 NA  71.0    NA 1.3 <NA>   NA <NA>         NA
#> 4852:    16.39 NA  51.5    NA 1.3 <NA>   NA <NA>         NA
#> 4853:    20.85 NA  65.5    NA 1.3 <NA>   NA <NA>         NA
#> 4854:    29.60 NA  93.0    NA 1.3 <NA>   NA <NA>         NA
#> 4855:    27.53 NA  86.5    NA 1.3 <NA>   NA <NA>         NA
#>       TreeFieldNumOriginal IdTreeOriginal CommercialSpOriginal DateOriginal
#>    1:                   45         100665                 TRUE   2016-09-14
#>    2:                  148         100765                FALSE   2016-09-14
#>    3:                   46         100666                 TRUE   2016-09-14
#>    4:                   49         100669                FALSE   2016-09-14
#>    5:                    1         100621                FALSE   2016-09-14
#>   ---                                                                      
#> 4851:                  922         101517                FALSE   2020-06-04
#> 4852:                  884         101479                FALSE   2020-06-04
#> 4853:                  887         101482                FALSE   2020-06-04
#> 4854:                  896         101491                 TRUE   2020-06-04
#> 4855:                  905         101500                FALSE   2020-06-04
#>       LifeStatusOriginal
#>    1:               TRUE
#>    2:               TRUE
#>    3:               TRUE
#>    4:               TRUE
#>    5:               TRUE
#>   ---                   
#> 4851:               TRUE
#> 4852:               TRUE
#> 4853:               TRUE
#> 4854:               TRUE
#> 4855:               TRUE

Rslt <- StatusCorrection(TestData[IdTree %in% selection])

StatusCorrectionPlot(Rslt)
#> [1] 1