UniqueMeasurement

Of the repeated measurements of an individual in the same census, keep only the measurements (rows) taken at the highest POM (KeepMeas = "MaxHOM"), and/or the most recent measurement (same census but more recent date) (KeepMeas = "MaxDate")

Usage

UniqueMeasurement(Data, KeepMeas = c("MaxHOM", "MaxDate"), ID = "IdStem")

Arguments

Data

Dataset (data.table)

KeepMeas

In case of multiple measurements in the same census: Possible values: "MaxHOM", "MaxDate" (character).

"MaxHOM": keep the measurement taken at the highest HOM/POM

"MaxDate": keep the most recent measurement (same census but more recent date)

ID

Column name indicating the identifier of the individual (character)

Value

Dataset (data.table) with 1 measurement (1 row) per IdCensus

Examples

library(data.table)

Data <- data.table( # 7 rows
IdStem = c(rep("a", 4), rep("b", 2), "c"),
IdCensus = c(rep(2000, 3), 2001, rep(2000, 3)),
Date = as.Date(c("2000-01-10", "2000-01-20", "2000-01-30", "2001-01-10",
                 rep("2000-01-10", 3))),
HOM = c(rep(1, 3), 2, 1, 2, 1)
)
Data
#>    IdStem IdCensus       Date HOM
#> 1:      a     2000 2000-01-10   1
#> 2:      a     2000 2000-01-20   1
#> 3:      a     2000 2000-01-30   1
#> 4:      a     2001 2001-01-10   2
#> 5:      b     2000 2000-01-10   1
#> 6:      b     2000 2000-01-10   2
#> 7:      c     2000 2000-01-10   1

UniqueMeasurement(TestData)
#>       MinDBH IdCensus Year Month Day       Date    Site Cluster Plot PlotViewID
#>    1:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    2:  10.03     2017 2017     7  19 2017-07-19 Paracou    <NA>    6       <NA>
#>    3:  10.03     2018 2018     7   8 2018-07-08 Paracou    <NA>    6       <NA>
#>    4:  10.03     2019 2019     6   7 2019-06-07 Paracou    <NA>    6       <NA>
#>    5:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#>   ---                                                                          
#> 4850:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4851:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4852:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4853:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 4854:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#>       Subplot PlotArea PlotElevation SubplotArea PlotLat PlotLon XPlotUTM
#>    1:       1     6.25            NA           0      NA      NA       NA
#>    2:       1     6.25            NA           0      NA      NA       NA
#>    3:       1     6.25            NA           0      NA      NA       NA
#>    4:       1     6.25            NA           0      NA      NA       NA
#>    5:       1     6.25            NA           0      NA      NA       NA
#>   ---                                                                    
#> 4850:       1     6.25            NA           0      NA      NA       NA
#> 4851:       1     6.25            NA           0      NA      NA       NA
#> 4852:       1     6.25            NA           0      NA      NA       NA
#> 4853:       1     6.25            NA           0      NA      NA       NA
#> 4854:       1     6.25            NA           0      NA      NA       NA
#>       YPlotUTM SubplotLat SubplotLon XSubplotUTM YSubplotUTM
#>    1:       NA         NA         NA          NA          NA
#>    2:       NA         NA         NA          NA          NA
#>    3:       NA         NA         NA          NA          NA
#>    4:       NA         NA         NA          NA          NA
#>    5:       NA         NA         NA          NA          NA
#>   ---                                                       
#> 4850:       NA         NA         NA          NA          NA
#> 4851:       NA         NA         NA          NA          NA
#> 4852:       NA         NA         NA          NA          NA
#> 4853:       NA         NA         NA          NA          NA
#> 4854:       NA         NA         NA          NA          NA
#>                   ScientificName      VernName        Family        Genus
#>    1:      Sandwithia guyanensis     wata tiki Euphorbiaceae   Sandwithia
#>    2:      Sandwithia guyanensis     wata tiki Euphorbiaceae   Sandwithia
#>    3:      Sandwithia guyanensis     wata tiki Euphorbiaceae   Sandwithia
#>    4:      Sandwithia guyanensis     wata tiki Euphorbiaceae   Sandwithia
#>    5:      Sandwithia guyanensis     wata tiki Euphorbiaceae   Sandwithia
#>   ---                                                                    
#> 4850:        Gustavia hexapetala  man tapuhupa Lecythidaceae     Gustavia
#> 4851:        Pradosia cochlearia       kimboto    Sapotaceae     Pradosia
#> 4852: Pogonophora schomburgkiana     geli koko      Peraceae  Pogonophora
#> 4853:        Pseudolmedia laevis satine rubane      Moraceae Pseudolmedia
#> 4854:          Inga lomatophylla    kodia weko      Fabaceae         Inga
#>              Species Subspecies Variety Voucher IdLevel Authority CommercialSp
#>    1:     guyanensis       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#>    2:     guyanensis       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#>    3:     guyanensis       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#>    4:     guyanensis       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#>    5:     guyanensis       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#>   ---                                                                         
#> 4850:     hexapetala       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#> 4851:     cochlearia       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#> 4852: schomburgkiana       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#> 4853:         laevis       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE
#> 4854:   lomatophylla       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#>       LifeForm TreeFieldNum IdTree StemFieldNum        IdStem TreeLat TreeLon
#>    1:     tree            1 100621         <NA> 100621_1_auto      NA      NA
#>    2:     tree            1 100621         <NA> 100621_1_auto      NA      NA
#>    3:     tree            1 100621         <NA> 100621_1_auto      NA      NA
#>    4:     tree            1 100621         <NA> 100621_1_auto      NA      NA
#>    5:     tree            1 100621         <NA> 100621_1_auto      NA      NA
#>   ---                                                                        
#> 4850:     tree         3008 418631         <NA> 418631_1_auto      NA      NA
#> 4851:     tree         3009 418632         <NA> 418632_1_auto      NA      NA
#> 4852:     tree         3010 418633         <NA> 418633_1_auto      NA      NA
#> 4853:     tree         3011 418634         <NA> 418634_1_auto      NA      NA
#> 4854:     tree         3012 418635         <NA> 418635_1_auto      NA      NA
#>       XTreeUTM YTreeUTM XTreePlot YTreePlot XTreeSubplot YTreeSubplot
#>    1:       NA       NA        NA        NA           NA           NA
#>    2:       NA       NA        NA        NA           NA           NA
#>    3:       NA       NA        NA        NA           NA           NA
#>    4:       NA       NA        NA        NA           NA           NA
#>    5:       NA       NA        NA        NA           NA           NA
#>   ---                                                                
#> 4850:       NA       NA        NA        NA           NA           NA
#> 4851:       NA       NA        NA        NA           NA           NA
#> 4852:       NA       NA        NA        NA           NA           NA
#> 4853:       NA       NA        NA        NA           NA           NA
#> 4854:       NA       NA        NA        NA           NA           NA
#>       LifeStatus Diameter BD Circ BCirc HOM  POM BHOM BPOM TreeHeight
#>    1:       TRUE    13.37 NA 42.0    NA 1.3 <NA>   NA <NA>         NA
#>    2:       TRUE    13.37 NA 42.0    NA 1.3 <NA>   NA <NA>         NA
#>    3:       TRUE    13.37 NA 42.0    NA 1.3 <NA>   NA <NA>         NA
#>    4:       TRUE    13.37 NA 42.0    NA 1.3 <NA>   NA <NA>         NA
#>    5:      FALSE    13.37 NA 42.0    NA 1.3 <NA>   NA <NA>         NA
#>   ---                                                                
#> 4850:       TRUE    10.19 NA 32.0    NA 1.3 <NA>   NA <NA>         NA
#> 4851:       TRUE    10.03 NA 31.5    NA 1.3 <NA>   NA <NA>         NA
#> 4852:       TRUE    10.03 NA 31.5    NA 1.3 <NA>   NA <NA>         NA
#> 4853:       TRUE    10.19 NA 32.0    NA 1.3 <NA>   NA <NA>         NA
#> 4854:       TRUE    10.19 NA 32.0    NA 1.3 <NA>   NA <NA>         NA
#>       TreeFieldNumOriginal IdTreeOriginal CommercialSpOriginal DateOriginal
#>    1:                    1         100621                FALSE   2016-09-14
#>    2:                    1         100621                FALSE   2017-07-19
#>    3:                    1         100621                FALSE   2018-07-08
#>    4:                    1         100621                FALSE   2019-06-07
#>    5:                    1         100621                FALSE   2020-06-04
#>   ---                                                                      
#> 4850:                 3008         418631                FALSE   2020-06-04
#> 4851:                 3009         418632                FALSE   2020-06-04
#> 4852:                 3010         418633                FALSE   2020-06-04
#> 4853:                 3011         418634                 TRUE   2020-06-04
#> 4854:                 3012         418635                FALSE   2020-06-04
#>       LifeStatusOriginal
#>    1:               TRUE
#>    2:               TRUE
#>    3:               TRUE
#>    4:               TRUE
#>    5:              FALSE
#>   ---                   
#> 4850:               TRUE
#> 4851:               TRUE
#> 4852:               TRUE
#> 4853:               TRUE
#> 4854:               TRUE